Added functionality to update member evaluation

+++ Server/CollaboRateAPIServer/CollaboRateAPIServer/Models/Rating.cs
@@ -9,7 +9,7 @@ namespace CollaboRateAPIServer.Models
         public int Group_ID { get; set; }
         public int Rater_ID { get; set; }
         public int Ratee_ID { get; set; }
-        public int Score {  get; set; }
+        public byte Score {  get; set; }
         public DateTime Rated_At { get; set; }
     }
 }
